Add FileOperationsService for core file actions
